refactor(architecture): extract contracts package
Freeze environment-free RPC, failure, and stream identifiers behind a workspace boundary before runtime extraction.
