+++ pnpm-workspace.yaml
@@ -1,6 +1,7 @@
 packages:
   - "."
   - "docs"
+  - "packages/*"
 
 minimumReleaseAge: 10080
 minimumReleaseAgeStrict: true
