+++ src/protocol/diagnostics-rpc.ts
@@ -1,7 +1,5 @@
+import { type RpcDefinition, RpcMethod } from "@ollama-client/contracts/rpc"
 import { z } from "zod"
-
-import type { RpcDefinition } from "./rpc"
-import { RpcMethod } from "./rpc"
 
 export const DiagnosticTestResultSchema = z
   .object({
