+++ src/protocol/extension-client.ts
@@ -1,7 +1,3 @@
-import { browser } from "@/lib/browser-api"
-import { createAppError, isAppError } from "@/lib/error-utils"
-
-import type { RpcRequest, RpcResponse } from "./provider-rpc"
 import {
   createRpcResponseEnvelopeSchema,
   RPC_CANCEL_MESSAGE_TYPE,
@@ -10,7 +6,10 @@ import {
   RpcErrorCode,
   type RpcMethod,
   type RpcRequestEnvelope
-} from "./rpc"
+} from "@ollama-client/contracts/rpc"
+import { browser } from "@/lib/browser-api"
+import { createAppError, isAppError } from "@/lib/error-utils"
+import type { RpcRequest, RpcResponse } from "./provider-rpc"
 import { RPC_METHOD_DEFINITIONS } from "./rpc-registry"
 
 const timeoutError = (method: RpcMethod) =>
